donation crud + rspec

+++ app/controllers/donations_controller.rb
@@ -1,13 +1,13 @@
 class DonationsController < ApplicationController
 
-  #include Pundit
-  #after_filter :verify_authorized
+  include Pundit
+  after_filter :verify_authorized
 
   def new
     @supporter = Supporter.find(params[:supporter_id])
     if @supporter
       @donation = @supporter.donations.build
-      #authorize @donation
+      authorize @donation
     else
       flash[:error] = 'Could not find supporter.'
       render root_path  # TODO: Fix. Render probably not correct
@@ -17,7 +17,7 @@ class DonationsController < ApplicationController
   def create
     @supporter = Supporter.find(params[:supporter_id])
     @donation = @supporter.donations.build(donation_params)
-    #authorize @donation
+    authorize @donation
     if @supporter.save
       flash[:success] = 'Success.'
       redirect_to supporter_path(@supporter)
@@ -30,21 +30,24 @@ class DonationsController < ApplicationController
     @donation = Donation.find(params[:id])
     @supporter = @donation.supporter
     @payments = @donation.payments
-    #authorize @donation
+    authorize @donation
   end
 
   def index
     @donations = Donation.all.paginate(:page => params[:page], per_page: 50)
+    authorize @donations
   end
 
   def edit
     @donation = Donation.find(params[:id])
     @supporter = @donation.supporter
+    authorize @donation
   end
 
   def update
     @donation = Donation.find(params[:id])
     @supporter = @donation.supporter
+    authorize @donation
     if @donation.update_attributes(donation_params)
       flash[:success] = 'donation updated.'
       redirect_to donation_path(@donation)
@@ -55,6 +58,7 @@ class DonationsController < ApplicationController
 
   def destroy
     donation = Donation.find(params[:id])
+    authorize donation
     donation.destroy
     flash[:success] = 'Donatoipn destroyed.'
     redirect_to donations_path
